+++ ps2redbook.h
@@ -8,7 +8,7 @@
 
 
 EXPORT void Redbook_XAStop(void);
-EXPORT int Redbook_XAPlay(int, int, int);
+EXPORT u8 Redbook_XAPlay(int, int, int);
 EXPORT u8 Redbook_XAPlayPos(int, int, CVector*, int);
 EXPORT void Redbook_XAAllow(bool);
 EXPORT void Redbook_XAInitAtStart(void);
